# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/admin/media:
    post:
      operationId: MediaController_upload
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                entityType:
                  type: string
                  example: dog
                entityId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaResponseDto'
        '401':
          description: ''
        '413':
          description: ''
      security: &ref_1
        - bearer: []
      summary: Upload image media
      tags: &ref_2
        - admin-media
  /api/v1/admin/media/{id}:
    get:
      operationId: MediaController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaResponseDto'
        '401':
          description: ''
        '404':
          description: ''
      security: *ref_1
      summary: Get media metadata and presigned URL
      tags: *ref_2
    delete:
      operationId: MediaController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
        '401':
          description: ''
        '403':
          description: ''
        '404':
          description: ''
      security: *ref_1
      summary: Soft-delete media
      tags: *ref_2
...
  /api/v1/found-reports:
    get:
      operationId: FoundReportsPublicController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            default: 20
            type: number
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - ACTIVE
              - HIDDEN
              - VERIFIED
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReportsPublicResponseDto'
      summary: List public found dog reports
      tags: &ref_6
        - found-reports
    post:
      operationId: FoundReportsPublicController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponseDto'
      summary: Submit a found dog report
      tags: *ref_6
  /api/v1/found-reports/{id}:
    get:
      operationId: FoundReportsPublicController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportPublicDetailDto'
      summary: Get public found dog report by id
      tags: *ref_6
  /api/v1/found-reports/{id}/media:
    post:
      operationId: FoundReportsPublicController_uploadMedia
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaResponseDto'
      tags: *ref_6
  /api/v1/admin/found-reports:
    get:
      operationId: FoundReportsAdminController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            default: 20
            type: number
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - ACTIVE
              - HIDDEN
              - VERIFIED
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReportsResponseDto'
      security: &ref_7
        - bearer: []
      summary: List found reports
      tags: &ref_8
        - admin-found-reports
  /api/v1/admin/found-reports/{id}:
    get:
      operationId: FoundReportsAdminController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponseDto'
      security: *ref_7
      summary: Get found report by id
      tags: *ref_8
    patch:
      operationId: FoundReportsAdminController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportStatusDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponseDto'
      security: *ref_7
      summary: Update found report status (hide, verify, restore)
      tags: *ref_8
  /api/v1/lost-reports:
    get:
      operationId: LostReportsPublicController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            default: 20
            type: number
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - ACTIVE
              - HIDDEN
              - VERIFIED
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReportsPublicResponseDto'
      summary: List public lost dog reports
      tags: &ref_9
        - lost-reports
    post:
      operationId: LostReportsPublicController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponseDto'
      summary: Submit a lost dog report
      tags: *ref_9
  /api/v1/lost-reports/{id}:
    get:
      operationId: LostReportsPublicController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportPublicDetailDto'
      summary: Get public lost dog report by id
      tags: *ref_9
  /api/v1/lost-reports/{id}/media:
    post:
      operationId: LostReportsPublicController_uploadMedia
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaResponseDto'
      tags: *ref_9
  /api/v1/admin/lost-reports:
    get:
      operationId: LostReportsAdminController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            default: 20
            type: number
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - ACTIVE
              - HIDDEN
              - VERIFIED
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReportsResponseDto'
      security: &ref_10
        - bearer: []
      summary: List lost reports
      tags: &ref_11
        - admin-lost-reports
  /api/v1/admin/lost-reports/{id}:
    get:
      operationId: LostReportsAdminController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponseDto'
      security: *ref_10
      summary: Get lost report by id
      tags: *ref_11
    patch:
      operationId: LostReportsAdminController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportStatusDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponseDto'
      security: *ref_10
      summary: Update lost report status (hide, verify, restore)
      tags: *ref_11
...
info:
  title: Dog Rescue API
  description: REST API for Dog Rescue (MVP scaffold)
  version: '1.0'
  contact: {}
...
servers:
  - url: /
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    MediaResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mimeType:
          type: string
          example: image/png
        sizeBytes:
          type: number
          example: 1024
        url:
          type: string
          description: Presigned GET URL (expires per MEDIA_PRESIGNED_TTL_SECONDS)
        createdAt:
          type: string
          example: '2026-08-02T00:00:00.000Z'
      required:
        - id
        - mimeType
        - sizeBytes
        - url
        - createdAt
...
    ReportPublicListItemDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        createdAt:
          type: string
        thumbnailUrl:
          type: object
        hasLocation:
          type: boolean
        verified:
          type: boolean
      required:
        - id
        - description
        - createdAt
        - hasLocation
        - verified
    PaginatedReportsPublicResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReportPublicListItemDto'
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - items
        - total
        - page
        - limit
    ReportPublicDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reporterName:
          type: string
        reporterPhone:
          type: string
        reporterEmail:
          type: object
        description:
          type: string
        latitude:
          type: object
        longitude:
          type: object
        createdAt:
          type: string
        verified:
          type: boolean
        media:
          type: array
          items:
            type: object
      required:
        - id
        - reporterName
        - reporterPhone
        - description
        - createdAt
        - verified
        - media
    CreateReportDto:
      type: object
      properties:
        reporterName:
          type: string
          example: Somchai
        reporterPhone:
          type: string
          example: '+66812345678'
        reporterEmail:
          type: string
          example: reporter@example.com
        description:
          type: string
          example: Small brown dog near the market.
        latitude:
          type: number
          example: 13.7563
        longitude:
          type: number
          example: 100.5018
        captchaToken:
          type: string
          description: Cloudflare Turnstile token
      required:
        - reporterName
        - reporterPhone
        - description
        - captchaToken
    ReportResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reporterName:
          type: string
        reporterPhone:
          type: string
        reporterEmail:
          type: object
        description:
          type: string
        latitude:
          type: object
        longitude:
          type: object
        status:
          type: string
          enum:
            - ACTIVE
            - HIDDEN
            - VERIFIED
        media:
          type: array
          items:
            $ref: '#/components/schemas/MediaResponseDto'
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - reporterName
        - reporterPhone
        - description
        - status
        - media
        - createdAt
        - updatedAt
    ReportListItemDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reporterName:
          type: string
        reporterPhone:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - HIDDEN
            - VERIFIED
        createdAt:
          type: string
        hasLocation:
          type: boolean
        verified:
          type: boolean
      required:
        - id
        - reporterName
        - reporterPhone
        - description
        - status
        - createdAt
        - hasLocation
        - verified
    PaginatedReportsResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReportListItemDto'
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - items
        - total
        - page
        - limit
    UpdateReportStatusDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - ACTIVE
            - HIDDEN
            - VERIFIED
      required:
        - status
...